windsurf-projects/context7-project/requirements.txt
pydantic-ai>=0.1.9
google-generativeai>=0.3.0
python-dotenv>=1.0.0
